+++ include/moss/core/key.hpp
@@ -1,42 +1,11 @@
-/*
- * @file    core/key.hpp
- * @brief   Singleton Contex class providing access to the ECS registry.
+/**
+ * @file    key.hpp
+ * @brief   Defines the `Key` class and permission system for ECS commands.
  *
- * This class provides a singleton instance for managing ECS operations, ensuring
- * that only one instance exists to prevent conflicting permission mappings.
- * It offers two primary modes of interaction:
- *
- * 1.  **Compile-Time Safe Commands:** Utilized through command structures, offering
- *     compile-time safety and type checking for ECS operations.
- *
- * 2.  **Runtime Fluent API:** A fluent design pattern (inspired by rs-bevvy) for
- *     runtime entity creation and component attachment. This approach, while
- *     sacrificing some compile-time safety, provides a more flexible and
- *     ergonomic API for rapid prototyping or less critical tasks.  The runtime
- *     API is accessed through the `Contex` functions (e.g., `create()`, `attach()`).
- *
- * The class uses mutex locks to guarantee thread-safe access and prevent
- * multiple instances from being created.
- *
- * **Usage Examples (Runtime API):**
- *
- * ```cpp
- * Contex::init()
- *     .create()
- *         .attach<PlayerTag, PlayerMovement, Position>()
- *     .create(10)
- *         .attach<EnemyTag, EnemyMovement, Position>();
- * ```
- *
- * **Singleton Management:**
- *
- * ```cpp
- * Contex::get();     // THROWS RUNTIME ERROR if not initialized
- * Contex::init();
- * Contex::get();     // Runs successfully
- * Contex::destroy();
- * Contex::get();     // THROWS RUNTIME ERROR after destruction
- * ```
+ * The `Key` class represents a permission system for commands in the ECS.
+ * It ensures that systems and commands only read or write to components 
+ * when permitted. The permissions are defined through the `Permissions` 
+ * enum and are used to control access to the ECS registry.
  */
 
 
@@ -52,13 +21,29 @@ namespace moss {
 
 namespace key {
 
-enum Permissions {
-    READ = 1 << 0,
-    WRITE = 1 << 1
-};
+    /**
+     * @enum Permissions
+     * @brief Defines permissions for command access to components.
+     *
+     * The `Permissions` enum provides two levels of access: `READ` and `WRITE`.
+     * These control the ability to read or modify components in the ECS registry.
+     */
+    enum Permissions {
+        READ = 1 << 0,
+        WRITE = 1 << 1
+    };
 
 }
 
+/**
+ * @class Key
+ * @brief Provides access control for ECS commands based on permissions.
+ *
+ * The `Key` class encapsulates the permissions for reading or writing to
+ * components in the ECS registry. It restricts access based on the 
+ * defined permissions (`READ` or `WRITE`). It is used by commands and
+ * systems.
+ */
 template<key::Permissions P>
 class Key {
 public:
